Updated openstack/openstack Project: openstack/ironic-python-agent  8173927dd3d11b8c9cf9afb6319ce39b1643a708
Move _get_agent_params() to a common place

The function _get_agent_params() parse the parameters passed to the agent
via kernel cmdline or vmedia. Other parts of the code needs to access
these parameters as well, so this patch is moving _get_agent_params()
and the related functions to a common place (utils.py).

Change-Id: I860f84d1d13511fff56d4aa56358ee597a9760d5