Updated openstack/openstack Project: openstack-infra/system-config  2b0a960ad3d97f6d9f04161bee6c61179ade187e
Enable logging of #openstack-nova

A while back, I resisted enabling logging of #openstack-nova.  I think
that was either wrong, or at least is no longer the better choice.
Logging is enabled for most channels and I no longer see a good reason
to not have that be the case for all OpenStack channels.

Change-Id: Id7ce058c3a4b21a3c700f16e17f388aa28229172