Updated openstack/openstack Project: openstack-infra/storyboard-webclient  e2b82ce1d0ebf1f93e9feab86e17d55967e3b5a7
Preference service uses private scope.

Using a 'this' reference when saving the preferences hash forces us
to manually manage scope, which is a bit of a pain. Switching over to
using a declared variable simplifies this. This patch also includes
a removal of angular.copy() as it was superfluous, and a call to $log
that provides runtime debug output on preference changes.

Change-Id: I682951403846b4b1fb6ed1e6fc82a4f5ba7c20b1