Updated openstack/openstack Project: openstack-infra/project-config  6d7360c98e522cef370500d8b831579063e51427
Ensure that gerritbot channels are in accessbot

Add new test that checks that channels mentioned in gerritbot are also
in accessbot.

Enhance the irc tox target so that it runs the checks, and update Zuul
so that the job is run for changes to gerritbot.

Add channels openstack-mistral and puppet-openstack to accessbot,
openstackinfra now has access to the channels. These two channels are
already in gerritbot.

Make check_irc_access.py script executable.

Change-Id: Ic68ba9a1c60854e68817fcadb1a5df2144944af7