Updated openstack/openstack Project: openstack/tempest-lib  e8be46fbf848d319793837c01b3bfb7d36c09227
Always print info logs from the rest client requests

The info logs make it very easy to parse the requests in tempest log
file. However on gate runs we run tempest with debug so we can see the
full req and resp for ever call. However it makes the log just a mess
of debug statements and difficult to trace through by hand. This commit
adds the info summary messages back even if debug is set so that we can
trace through. There will be a slight duplication of logging, but it is
worth the tradoff.

Change-Id: I5499f231a91342a184d254e95bc72a7480a0ece0
Depends-On: Ia0de957b681cb924a57af98d99a9389ee234ed5b