Updated openstack/openstack Project: openstack/python-novaclient  3db98546cff3139f07d2d4a7172483ba3323b6fc
nova flavor-show command is inconsistent

The nova flavor-show command accepts a flavor
name as its parameter. The command show inconsistency
as it works with the exact flavor name and flavor name
in all lowercase but fails for any other case pattern.

So, updated the code so that if we run the command
with flavor-name case patern it runs fine.

Closes-Bug: #1423885
Change-Id: I713bf2b5aca977df40dc745e29d4fe98a950c991