Updated openstack/openstack Project: openstack/nova  243932315475115df713a45ebbe31e850181a187
Add instance_mappings table to api database

This creates the instance_mappings table for cells v2.  This table will
contain a mapping for which cell an instance belongs to.  project_id is
included so that it is easy to get a full listing of instances for a
project.

Change-Id: Ic5a2f0add7455ff030fa8acc41b9304205ced715
bp: cells-instance-mapping