Updated openstack/openstack Project: openstack/nova  23f4f556f5575361063d8cde3c0a60e5e9d2652c
Refactor _regex_instance_filter for testing

_regex_instance_filter was determing which regex operation to use based
on the database connection config.  This led to writing tests that
modified the connection config to be different than the one which
created the engine.  This was mostly harmless but now that we're looking
to add the ability to get different engines based on that config value
this can lead to confusion and errors.

There was a test class intended to test this behavior but only a few of
the tests in there actually did.  The class has renamed more
appropriately and tests which did not belong were moved to other test
classes.

Change-Id: I06749f5f6e2ce9bf28c378a022f862fe2db3fcb5