Updated openstack/openstack Project: openstack/nova  646c45958da26b81f5db2b92ef98686ec6f868a2
Ban passing contexts to remotable methods

We've been wanting to do this for a while and have mostly cleaned up the
entire code base. This bans passing context to remotable methods by force
to prevent any new cases from coming in. When we move to o.vo, remotable
methods *cannot* take a context, and so this wedge (and the method signatures
of remotables) will be removed and cleaned up.

Related to blueprint kilo-objects

Change-Id: I828e9a34fa62f84991533e74e9685bb391ca4a3c