Updated openstack/openstack Project: openstack/api-site  cf9c7c0f4cd6cb6c3e182ee9487ebc6488915e1f
Changing incorrect GET server details for Compute

The GET details for /servers and /servers/details currently shows
request parameters for image and flavor as Name of image/flavor in
URL and is of xsd:anyURI type. Change these parameters to be of type
csapi:UUID and edit the description accordingly.

Change-Id: I98ad0c5175fe1f9dcbe834ddf2944d9b5c66f055
Closes-Bug: #1426795