Updated openstack/openstack Project: openstack/horizon  a69ab4bf767d393680d59fd33ae528808e63e0aa
Angular-based Modal Wait Spinner Widget

When all UI action must be blocked in Horizon, such as waiting for
a new page, or submitting a server request, Horizon shows a wait
spinner. The current spinner does not correctly layer on top of
Angular bootstrap modal components (such as the wizard).

This widget has the same look and feel as the current Horizon
spinner, but provides an Angular service to allow easy control
from Angular components, and correctly layers on top of any
Angular bootstrap modals that are currently shown.

Currently, this is intended as a "global" wait spinner. It does
not yet support being used as a wait spinner within other
components (such as waiting for a cell of data to load within
a table).

Partially Implements: blueprint launch-instance-redesign

Change-Id: I1e93ee2e2afde8cc55daa42148dfcdb4d0496d89