Updated openstack/openstack Project: openstack/cinder  67896cc81732656ee7313e27cce4b4e24a150927
Add multiattach capabilities injection

This patch is a follow up to the multiattach patch that adds
the ability to the scheduler to be able to filter capabilities
on the volume multiattach flag.   When a volume creation request
comes in with the multiattach flag set on a volume, there was no
way to filter out backends that don't support attaching a volume
to more than one host.   By defaut, no host supports it.  When
The Nova side lands, we can adjust drivers to add the new
multiattach:True capability reporting.

Change-Id: Ia426a2bf2ab8214f52fb6e70a3a1deff68d13d98