Updated openstack/openstack Project: openstack/manila  68c5f2c89b2633a7444fdb8aa6c1c9205e51007e
Always allow delete share-network when no shares exist

Currently tenants are unable to delete share_networks if there are existing
share_servers. This doesn't make sense, because the share_server exists only
for the share_network, and if the tenant is done with the share_network,
then the share_server can go away safely.

Change-Id: I8925f3deba64da6b011516163e5554b8134c8c0a
Closes-Bug: #1429123