Updated openstack/openstack Project: openstack/neutron  bed4d80b7e54764104bde7cce4a3bfb1c4499e90
Raise QuotaResourceUnknown in the quota engine

This patch lets the quota engine verify whether it manages a resource
for which a limit is being checked.

So far this check has been delegated to the drivers. This is
conceptually wrong and also lead to code duplication.

Unit tests and some docstrings are also fixed accordingly as a
part of this patch.

Closes-Bug: #1430519
Related to blueprint better-quotas

Change-Id: If1467881f13e30afa53a23b904f8eae5c7264834