Updated openstack/openstack Project: openstack-infra/project-config  0927a71eaf96ec1f2392d89141edb65fc390be81
Add option to set swift delete after header

Set the swift delete after header in the zuul swift upload script so
that swift handles log deletion for us. Default to a delete after of 6
months but allow the value to be configurable.

Headers are sent per a post. Posts are done per file-job (or queue
entry). Additionals headers can be set up when building the queue.

Co-Authored-By: Joshua Hesketh <josh@nitrotech.org>
Change-Id: Idd3b4b0c9a6819cda5e9b36ce188a61b558873be