Updated openstack/openstack Project: openstack-infra/project-config  cd37d1a8de09013e8b9817a3ce026ce82b1ba361
Add skip if for tempest running tempest on tempest commits

This commit adds a skipIf check for running tempest jobs on specific
tempest commits. If the proposed commit is a doc only, or a sample
config only change there is no need to run the tempest tests on the
commit.

Change-Id: I664288625727a19edcba6b19343d4dd9762f86dd