Updated openstack/openstack Project: openstack/heat  3f36398e670277f2b5ca392b2bae5b21a6301c6c
Use AWS::EC2::EIP for test_server_cfn_init

Many (most) clouds won't have the assumed network to directly ping/ssh
to the default IP address. This changes test_server_cfn_init to assign
a floating IP, and use that to ping and ssh to the server.

It also moves the ping to the end of the test. This was initially
meant to catch regressions in a co-gated neutron, but it is no longer
required for a heat-only test.

Change-Id: Ia9404289b2f9632883a479ccb435cbf6c51fcf2f