Updated openstack/openstack Project: openstack/cinder  06d97fc16e8c5eff9ed12889fda9cebcb91474d6
Unset auth token before trying to login to Datera

After a token expires, the Datera driver would try to login, and pass
along the expired token. This would fail because the receiving server
will authorize anytime it's given an auth token, even for endpoints that
don't require it. This change unsets the auth token when logging in, so
it's not passed a long.

Closes-Bug: #1430978
Change-Id: Id445718e3e73b676e3af8005b8fdd1a028ef2b0e