Updated openstack/openstack Project: openstack/cinder  1dda708581d4c2cdf073ab290356eab16112300f
Update hacking check for oslo_log

With commit 5980da5d we moved Cinder to using the olso.log
library.  We also moved to the new oslo_log namespace.  With that
change we should have also updated the hacking checks to include
looking for old uses of oslo.log.

This change adds the necessary hacking check.

Change-Id: I6a5d4b66d3258dbd01543be03810a8f84040e660