Updated openstack/openstack Project: openstack/nova  1d0cf60f3c2be68d8c90d70b2f2211d6b54af4f5
Remove useless policy rules for v2.1 api which removed/disabled

Because V2.1 must be identical with V2 and must not contain any new
attributes and new method. So the swap/attach/deatch in extend_volume
was removed.  But there still left some policy rules related to it.
This patch cleanup them.

Change-Id: I6e7e3263261519f72fa6b067db6a5930026f02dd