Updated openstack/openstack Project: openstack/python-tuskarclient  284c47b0c8582dd8d5fb2e8d7974513d88d9eb40
Updated URLs to remove version string

The keystone endpoint was changed to include the v2 portion of the URL,
which the client previous added itself. This resulted in URLs containing
the v2 part twice. This fix removes the version string from the URLs the
client uses to connect to Tuskar.

Change-Id: I3f20ab631770dfb244c0480484d3b4c19736a242