Updated openstack/openstack Project: openstack/designate  445ebf9604b5fae3d12edd12e24e1703e4ce01d6
Create DesignateAdapter and surrounding infrastructure

This implements the basic infrastructure for DesignateAdapter
and creates the basis for the API v2/v1 Adapter objects

These will replace views in both the v1 and v2 APIs, and allow for the
validation of API requests to move into the DesignateObjects themselves.

Example use is

   zone = central.get_domain(blah)
   DesignateAdapter.render('API_v1', zone)
   // Returns v1 dict, ready to	convert to JSON

These objects have a registry of sorts, so all calls will be made to
DesignateAdapter, and the combination of the format param ('API_v1' in
this case) will load the right Adapter and return the right JSON ready
dict.

Change-Id: I2b77205751675de600248180fafbfe22a2c1d8f5
Blueprint: validation-cleanup