Updated openstack/openstack Project: openstack/taskflow  c759d0e5a230a3c70236abd6136d6f21db38c980
Use ordered[set/dict] to retain ordering

Instead of using always using a set/dict which do not retain
use a ordered set and a ordered dict for requires, optional,
and provides and rebind mappings types so that the ordering
of these containers is maintained later when they are used.

These ordering can be useful depending on the atom type (such
as in a map and reduce tasks).

Partial-Bug: 1357117

Change-Id: I365d11bbba4aa221bc36ca15441acecf199b4d56