Updated openstack/openstack Project: openstack/keystone  6d72d30c12f43f091a43894d7a4b5ebe98bc74b1
Adding utf8 to federation tables

Tables have to be in utf8 for the migration utility
to work properly. This fixes the UTF8 issues with the
federation tables in-line and should be able to be
backported. A subsequent change to handle sanity checks
in a more sane way (post migrations) will be implemented
going forward.

Change-Id: Id68c3f307f53b6a4de011e2608d899d044aced68
Closes-Bug: 1426334