Updated openstack/openstack Project: openstack/keystone  6ebada6efab662a3558751d074d45feac27faf0e
Migrations squash

SQL Migrations for Kilo require the schema version to be at 044
which is the final verison for the Icehouse release. This limits the
change sets that could affect assumptions made in earlier migration
files. Migrating to before the Icehouse release migration is no longer
supported.

Implements: blueprint kilo-sql-squash

Change-Id: I65c48f1dc336cef14a365e37dc4084a4abc6ee51