Updated openstack/openstack Project: openstack/openstack-manuals  ba2dc5aa5e267fc6b724b17dd61bf4334919078d
Adds hypervisor_version_requires (ImagePropertiesFilter)

This is to prevent instances with newer tools from being
provisioned on older hypervisors.

Added comparison of hypervisor version from the
compute-node and image metadata in the image properties
filter.

In host scheduler image_prop_filter, the hypervisor_version_requires
image properties, if available, is compared to the host_state's
hypervisor_version, if available.

hypervisor_version_requires should be added on the image
as metadata and should have a operator and version value.
Examples : ">=6.0", ">6.0, <6.2", ""!=6.1".

Change-Id: I77dce6a5c28b5a25766b0273dc58a9dbb08defdd
Closes-bug: 1246231