Updated openstack/openstack Project: openstack/neutron  c6e54613d96f5e9c11b405750e372e4cf4f429dc
Set default of api_workers to number of CPUs

The current default for api_workers is 0.  This patch changes it to
use the number of CPUs available, which is the default that will
provide best performance.  This is consistent with some other
OpenStack service (Nova and Swift, at least, perhaps others).

Note that devstack is updated in another patch to ensure multiple API
workers is being tested for Neutron the same way it's tested for
several other services:

    https://review.openstack.org/140482

Co-Authored-By: Eugene Nikanorov <enikanorov@mirantis.com>
DocImpact
UpgradeImpact
Change-Id: I4d76baf68b014fc80693f53874734783089abda2