Updated openstack/openstack Project: openstack/neutron-vpnaas  5de38ff18e4caf402c0e337e084219a100b6e466
Fix functional test breakage from DevStack change

The commit bf2ad70 [1] for DevStack, moved lib/config to inc/meta-config
and this breaks the to VPN functional jobs, which source this file in
the gate_hook.sh script.

This is a quick fix to get the jobs working again. A future commit can
consider updating gate_hook.sh (as I see that Neutron repo is using a
different script now).

[1] https://review.openstack.org/#/c/162784/

Change-Id: I298b302ea4860e23acad10087a4ebe0783780826