Updated openstack/openstack Project: openstack-infra/puppet-zuul  9b4a4cfcfe8da9caf8b171b31720f4b6107d123d
Option to allow HTTPS for the proxy

If you want to use Javascript to include status.json within the
context of a page served via HTTPS, browsers will basically insist
that the status.json be served via HTTPS as well. This patch
provides an option to add HTTPS for the Apache proxy vhost if
desired.

Change-Id: I9799f39bf170f660bcbc17719937e1e87b68ac4a