Updated openstack/openstack Project: openstack/tempest  06af51c2f03699b3663bf44189992a1271795f8a
Create test nodes for test_baremetal_nodes

This test currently relies on pre-existing resources and races if run
in parallel to other baremetal tests.  This adds creation of 3 test
nodes directly in Ironic to be tested via the Nova API extension.

This also tags said test with the 'baremetal' test attribute.

Closes-bug: #1425258

Change-Id: I4dbd37bdb2019b6eb0140d46a605d5c8392323f4