Updated openstack/openstack Project: openstack/nova  0d0360907fbf81a919c6803659e1ab2963e5c474
Remove usage of remotable context parameter in ec2 object

This removes the use of the context parameter in remotable methods
in the ec2 object. This is a precursor to actually removing the
context from the call, but we need to remove dependence on the
argument first. Remotable methods should be using the bundled object
context.

Related to blueprint kilo-objects

Change-Id: I135dcef64320dc186ea555fb5f3cb982504f87f9