Updated openstack/openstack Project: openstack/horizon  b6ccb8eb4d30ff689dbf4d832d3d7db3f0d1089b
Updated transfer tables

The updated transfer tables now provide a warning
popover that appears if the user clicks a disabled
allocated button. The popover contains a message and
link to expand the details for the row. Popover should
dismiss on click elsewhere. A warning icon now shows
on the bottom right of the disabled button.

This patch also includes:
  - cleanup allocate section to hide count for single allocation
  - cleanup of directive code and documentation
  - cleanup of table CSS and transfer table CSS
  - moved re-order column to first column (if applicable)
  - addition of IrDragNDrop to hz-table
  - updated tests

Partially Implements: blueprint launch-instance-redesign

Change-Id: Icf2046a9e620a8a0efb2bdc7a7ddb18f6445a20f