Updated openstack/openstack Project: openstack/keystone  24bc6a1bf03e0ef71b16b2e973120aa9a8131778
Rename get_events to list_events on the Revoke API

Methods that return a list of entities are named
list_<entity-name>s throughout the project. For
example: list_projects and list_users.

In the Revoke API, the method to list revocation
events was called get_events(). This patch
renames it to list_events(), in order to keep
consistency.

Change-Id: I1032916201a46476b92b44aa7fbcffdcb072f6ea