Updated openstack/openstack Project: openstack-infra/puppet-subunit2sql  a665f7b9709fd3bca1ebfc7b7bd8e8dd4cae8a28
Set timeouts for db upgrade operations

This commit is adding timeout fields to the 2 execs as part of a new
subunit2sql release. The first sets the timeout for the mysqldump to
be 30min which should be sufficient time for that. The timeout for the
actual db upgrade command is disabled however, because there is no
way to safely estimate how long the migration will take.

Change-Id: I7deb7abca9c97186343eb6c3ea30185c23c24c7a