Updated openstack/openstack Project: openstack/nova  2d799f932cc53f40973e30668d83c25e0e257fa3
Make service-update work in API cells

nova service-update is handled by HostAPI.service_update method.
Normally, it searches for relevant Service object in DB and calls
update with parameters-to-update on it. However, in API cell it is
over-ridden so that service_update message can be routed to appropriate
cell. Once the destination cell is reached, the message processor
invokes HostAPI.service_update to perform the actual update. This
works in case of child cells.
However in case of API cells, since HostAPI.service_update has been
over-ridden, when the message processor invokes HostAPI.service_update
to perform actual update of the service, the API cell tries to again
route the message and fails in doing so with an AttributeError (as the
message is already at the destination cell and can not be further
routed).
This patch moves the object update() and save() call to a separate
method and modifies the message processor to call the new method. The
original service_update is also modified to call the new method.

Change-Id: I1da32d83e0decfa7240a85bbd45354d391bba021
Closes-Bug: 1398349