Updated openstack/openstack Project: openstack/cinder  0747b83a395a4e3c657e010bc30c44ec9d78d2bd
Fix LVM thin pool creation race

In the event that two copied of the LVM driver get init called at
the same time (e.g. cinder-volume and cinder-backup getting
started in parallel, on the same host), it is possible for the
thin pool check/create to race. Add a simple recheck if the create
fails, to cover this window.

Change-Id: I006970736ba0e62df383bacc79b5754dea2e9a3e
Closes-Bug: #1410341