Updated openstack/openstack Project: openstack/taskflow  dbb71b8471c459b4631da7a40ed5987b28ec2ab5
Allow passing 'many_handler' to fetch_all function

When many values are found for a given result a callback
is called to determine what to return. To match the fetch
routine allow for the fetch_all function to pass along a
handler (and default to one that is the existing behavior)
so that people can extract there own values (if they so
choose).

Change-Id: I8c105e1a810e8e0c2c210613506c9ce59021009d