Updated openstack/openstack Project: openstack-infra/project-config  a469266dd145fb5cebb57e3e463fd7b9a66b89da
Cap the subunit2sql version in the nodepool scripts

Subunit2sql 0.4.0 was recently released which included a schema change
however the infra db migration takes too long and can't be automated.
(yet?) Until we run the migration we can't use 0.4.0 on any of the
tooling. This caps the version to be less than 0.4.0

Change-Id: I08f113fa904fa962e8f2dc04187ff44e764de47e